requirements.txt
streamlit>=1.32.0
openpyxl>=3.1.2
thefuzz>=0.22.1
python-Levenshtein>=0.25.0
python-docx>=1.1.0
pdfplumber>=0.11.0
requests>=2.31.0
python-dotenv>=1.0.0
